Stream avaragePrice changes

+++ src/OOP/SoundCard.java
@@ -1,7 +1,7 @@
 package OOP;
 
 public class SoundCard extends Motherboard{
-    public SoundCard(int id, String producer, String model, String serialNumber, ConnectionType connectionType, long price) {
+    public SoundCard(int id, String producer, String model, String serialNumber, ConnectionType connectionType, double price) {
         super(id, producer, model, serialNumber, connectionType, price);
     }
